pyproject.toml
[project]
name = "docs4llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "html2text>=2025.4.15",
    "pytest-playwright>=0.7.0",
    "requests>=2.32.3",
]
